fix issue dio service

+++ lib/models/weather_hourly_forecasrs.dart
@@ -57,7 +57,7 @@ class WeatherHourlyForecasrs {
 }
 
 class Temperature {
-  int? value;
+  double? value;
   String? unit;
   int? unitType;
